Moving towards #149. Project structure changed a bit.
