+++ libakumuli/storage_engine/nbtree.h
@@ -77,71 +77,15 @@
 #include <deque>
 
 // App headers
+#include "nbtree_def.h"
 #include "blockstore.h"
 #include "compression.h"
+#include "operators/operator.h"
 
 
 namespace Akumuli {
 namespace StorageEngine {
 
-
-enum class NBTreeBlockType {
-    LEAF,   // data block
-    INNER,  // super block
-};
-
-
-enum {
-    AKU_NBTREE_FANOUT = 32,
-};
-
-
-/** Reference to tree node.
-  * Ref contains some metadata: version, level, payload_size, id.
-  * This metadata corresponds to the current node.
-  * Also, reference contains some aggregates: count, begin, end, min, max, sum.
-  * This aggregates corresponds to the current node if leve=0 (current node is a
-  * leaf node) or to the pointee if level > 0. If level is 1 then pointee is a
-  * leafa node and all this fields describes this leaf node. If level is 2 or more
-  * then all this aggregates comes from entire subtree (e.g. min is a minimal value
-  * in leaf nodes in pointee subtree).
-  */
-struct SubtreeRef {
-    //! Number of elements in the subtree
-    u64 count;
-    //! Series Id
-    aku_ParamId id;
-    //! First element's timestamp
-    aku_Timestamp begin;
-    //! Last element's timestamp
-    aku_Timestamp end;
-    //! Object addr in blockstore
-    LogicAddr addr;
-    //! Smalles value
-    double min;
-    //! Registration time of the smallest value
-    aku_Timestamp min_time;
-    //! Largest value
-    double max;
-    //! Registration time of the largest value
-    aku_Timestamp max_time;
-    //! Summ of all elements in subtree
-    double sum;
-    //! First value in subtree
-    double first;
-    //! Last value in subtree
-    double last;
-    //! Node version
-    u16 version;
-    //! Node level in the tree
-    u16 level;
-    //! Payload size (real)
-    u16 payload_size;
-    //! Fan out index of the element (current)
-    u16 fanout_index;
-    //! Checksum of the block (not used for links to child nodes)
-    u32 checksum;
-} __attribute__((packed));
 
 //! Result of the aggregation operation that has several components.
 struct NBTreeAggregationResult {
@@ -194,15 +138,15 @@ struct NBTreeCandlestickHint {
 };
 
 
-/** NBTree iterator.
+/** Database query operator.
   * @note all ranges is semi-open. This means that if we're
-  *       reading data from A to B, iterator should return
+  *       reading data from A to B, operator should return
   *       data in range [A, B), and B timestamp should be
   *       greater (or less if we're reading data in backward
   *       direction) then all timestamps that we've read before.
   */
 template <class TValue>
-struct NBTreeIteratorBase {
+struct QueryOperator {
 
     //! Iteration direction
     enum class Direction {
@@ -210,9 +154,9 @@ struct NBTreeIteratorBase {
     };
 
     //! D-tor
-    virtual ~NBTreeIteratorBase() = default;
+    virtual ~QueryOperator() = default;
 
-    /** Read data from iterator.
+    /** Read next portion of data.
       * @param destts Timestamps destination buffer. On success timestamps will be written here.
       * @param destval Values destination buffer.
       * @param size Size of the  destts and destval buffers (should be the same).
@@ -224,10 +168,10 @@ struct NBTreeIteratorBase {
 };
 
 //! Base class for all raw data iterators.
-using NBTreeIterator = NBTreeIteratorBase<double>;
+using NBTreeIterator = QueryOperator<double>;
 
 //! Base class for all aggregating iterators. Return single value.
-using NBTreeAggregator = NBTreeIteratorBase<NBTreeAggregationResult>;
+using NBTreeAggregator = QueryOperator<NBTreeAggregationResult>;
 
 
 /** NBTree leaf node. Supports append operation.
