+++ libakumuli/storage_engine/operators/operator.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace Akumuli {
+namespace StorageEngine {
+
+}
+}
